old/requirements.txt
fastapi
uvicorn
httpx
PyJWT
python-dotenv
pydantic
